updated login files so that users can login and logout
When the app launches if a user is already signed in, it launches into
their view. If they are not it prompts them to login. If the user does
not already exist the user is created. Also modified the login option
previously given in options menu to logout, this functions correctly
and prompts the user to login.